fix: mv async ServeOptions to options.py and delete async.types.ts
